refactor(vtc): shared DID-VM resolver + issuer-binding helper (P2.4 part 1)
Three call sites hand-rolled "resolve a DID doc -> find the verificationMethod
-> pull the key -> verify a Data-Integrity proof". The credential-exchange path
already did it the right way — it delegates resolution to the
affinidi-data-integrity library's `DataIntegrityProof::verify`, handing it a
`VerificationMethodResolver`. This hoists that one resolver so the others verify
through the same library path instead of re-implementing key resolution.

- New `credentials::vm_resolver`: `DidVmResolver` (the single
  `VerificationMethodResolver`, moved out of `exchange/verify.rs`) +
  `check_issuer_binding(vm, issuer)` — the "the proof's key must sit under the
  declared issuer" check the three paths each inlined.
- The shared resolver's Ed25519 extraction now uses the upstream
  `VerificationMethod::get_public_key_bytes()` (Multikey + Ed25519Verification
  Key2020 + JWK) instead of exchange's old `publicKeyMultibase`-only decode — a
  strict superset, so no key shape is dropped. The `did:key` fast path and the
  BBS-G2 path are unchanged.
- `exchange/verify.rs` now imports the shared resolver and uses
  `check_issuer_binding` for its DI + bbs-2023 proof-binding checks (the SD-JWT
  JWS-`kid` binding, a different mechanism, stays inline).
- `routes/relationships.rs::verify_vc_proof` drops its hand-rolled resolution
  and verifies via `proof.verify(&vrc, &DidVmResolver::new(Some(resolver)),
  opts)` + `check_issuer_binding`.

No behaviour change (exchange + relationships suites green), no wire/public-API
change. Recognition's bespoke `ForeignIssuerKeyResolver` + its DidResolver copy
are converted in part 2 (it carries the test-stub rework).

`cargo test -p vtc-service --lib` → 580 passed; relationships → 10 passed; fmt +
clippy clean.
